+++ src/break.h
@@ -46,10 +46,10 @@ void SortBreakInfo(void);
 BREAK_INFOp SetupWallForBreak(WALLp wallp);
 BREAK_INFOp SetupSpriteForBreak(SPRITEp sp);
 short FindBreakSpriteMatch(short match);
-BOOL HitBreakWall(WALLp wp, long, long, long, short ang, short type);
+BOOL HitBreakWall(WALLp wp, int, int, int, short ang, short type);
 int HitBreakSprite(short BreakSprite, short type);
 BOOL CheckBreakToughness(BREAK_INFOp break_info, short ID);
-int WallBreakPosition(short hitwall, short *sectnum, long *x, long *y, long *z, short *ang);
+int WallBreakPosition(short hitwall, short *sectnum, int *x, int *y, int *z, short *ang);
 void SortBreakInfo(void );
 
 #endif
